refactor: extract ActivityMapPanel from FitnessStatusDetail
